refactor: Strategy関連クラスの配置を整理
